+++ src/test/java/HeuristicTest.java
@@ -6,8 +6,8 @@ public class HeuristicTest {
 
     @Test
     public void eucildeanDistanceTest() {
-        int[][] table = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}, {13, 14, 15, 0}};
-        int[][] table2 = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 0}, {13, 14, 15, 12}};
+        byte[][] table = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}, {13, 14, 15, 0}};
+        byte[][] table2 = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 0}, {13, 14, 15, 12}};
         State state = new State(table);
         State copy = new State(table2);
         Heuristic heuristic = new EuclideanDistance();
@@ -16,8 +16,8 @@ public class HeuristicTest {
     }
     @Test
     public void manhatanDistanceTest() {
-        int[][] table = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}, {13, 14, 15, 0}};
-        int[][] table2 = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 0}, {13, 14, 15, 12}};
+        byte[][] table = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}, {13, 14, 15, 0}};
+        byte[][] table2 = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 0}, {13, 14, 15, 12}};
         State state = new State(table);
         State copy = new State(table2);
         Heuristic heuristic = new MannhatanDistance();
@@ -26,8 +26,8 @@ public class HeuristicTest {
     }
     @Test
     public void hammingDistanceTest() {
-        int[][] table = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}, {13, 14, 15, 0}};
-        int[][] table2 = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 0}, {13, 14, 15, 12}};
+        byte[][] table = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}, {13, 14, 15, 0}};
+        byte[][] table2 = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 0}, {13, 14, 15, 12}};
         State state = new State(table);
         State copy = new State(table2);
         Heuristic heuristic = new MannhatanDistance();
